// src/app/shared/models/question-details.interface.ts
export interface QuestionDetail {
    paper_id: string;
    grade: string;
    subject_id: string;
    published_year: string;
    published_month: string;
    paper_no: string;
    question_no: string;
    marks: number;
    qno: number;
    atp_id: string;
    topic: string;
    week: string;
    term: string;
    duration: number;
}